requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.1
numpy>=1.24.0
langchain>=0.1.12
langchain-community>=0.0.28
docarray==0.40.0 